feat(homepage): update components and styles
