+++ src/App.css
@@ -39,17 +39,17 @@ html, body, #root {
     @apply py-2 px-4 font-bold rounded transition-transform transform active:scale-95;
   }
   
-  /* Primary Button Variant */
-  /* .btn-primary { */
-    /* @apply bg-white text-customPink border border-customPink  */
-      /* hover:bg-customRed hover:text-white  */
-      /* focus:outline-none focus:ring-2 focus:ring-customPink; */
-  /* } */
+  Primary Button Variant
+  .btn-primary {
+    @apply bg-white text-customPink border border-customPink 
+      hover:bg-customRed hover:text-white 
+      focus:outline-none focus:ring-2 focus:ring-customPink;
+  }
   
-  /* Secondary Button Variant (optional) */
-  /* .btn-secondary { */
-    /* @apply bg-gray-500 text-white hover:bg-gray-600 focus:ring-gray-400 focus:outline-none; */
-  /* } */
+  Secondary Button Variant (optional)
+  .btn-secondary {
+    @apply bg-gray-500 text-white hover:bg-gray-600 focus:ring-gray-400 focus:outline-none;
+  }
 
   .link {
     @apply text-white no-underline relative inline-block transition-all duration-300 ease-in-out text-xl;
