+++ src/components/home/navbar.component.tsx
@@ -1,7 +1,8 @@
 import { useState } from "react";
 
+
 const NavBarComponent: React.FC = () => {
-  const Links = ["Home", "About", "Contact"];
+  const Links = ["Home", "How To Rent", "Choose Us", "About", "Contact"];
   const [isOpen, setIsOpen] = useState(false);
   const [user, setUser] = useState<null | { name: string; avatar: string }>(null);
   const [dropdownOpen, setDropdownOpen] = useState(false);
@@ -24,22 +25,21 @@ const NavBarComponent: React.FC = () => {
   const toggleMenu = () => setIsOpen(!isOpen);
 
   return (
-    <nav className="bg-[#B9C2C8] shadow-md top-0 sticky w-full z-100">
-      <div className="container mx-auto px-4 flex justify-between items-center h-20">
+    <nav className="bg-[#970747] shadow-md top-0 sticky w-full z-50">
+      <div className="container mx-auto px-4 flex justify-between items-center h-16">
         {/* Brand */}
-        {/* Logo */}
         <img
-          src="/images/logo3.png"
-          alt="EasyDrive Logo"
-          style={{ width: "170px", objectFit: "contain" }}
-        />
+            src="/images/newlogo.png"
+            alt="EasyDrive Logo"
+            style={{ width: "120px", objectFit: "contain" }}
+          />
         {/* Navigation Links */}
         <div className="hidden md:flex space-x-8">
           {Links.map((link) => (
             <a
               key={link}
               href={`/${link.toLowerCase()}`}
-              className="#0d5f07 hover:text-gray-300"
+              className="text-white hover:text-gray-300"
             >
               {link}
             </a>
@@ -48,6 +48,8 @@ const NavBarComponent: React.FC = () => {
 
         {/* Hamburger Menu, Avatar, and Logout */}
         <div className="flex items-center space-x-4 md:hidden">
+
+
           {user ? (
             <>
               <img
@@ -81,7 +83,7 @@ const NavBarComponent: React.FC = () => {
               )}
             </>
           ) : null}
-          <button
+                    <button
             className="text-white hover:text-gray-300"
             onClick={toggleMenu}
           >
@@ -143,12 +145,12 @@ const NavBarComponent: React.FC = () => {
           ) : (
             <>
               <button
-                className="#0d5f07 border border-white px-4 py-2 rounded hover:bg-white hover:text-[#526877]"
+                className="text-white border border-white px-4 py-2 rounded hover:bg-white hover:text-[#970747]"
                 onClick={handleLogin}
               >
                 Sign In
               </button>
-              <button className="bg-white text-[#526877] px-4 py-2 rounded hover:bg-gray-200">
+              <button className="bg-white text-[#970747] px-4 py-2 rounded hover:bg-gray-200">
                 Sign Up
               </button>
             </>
@@ -163,7 +165,7 @@ const NavBarComponent: React.FC = () => {
             <a
               key={link}
               href={`/${link.toLowerCase()}`}
-              className="block text-gray-800 px-4 py-2 hover:text-[#526877]"
+              className="block text-gray-800 px-4 py-2 hover:text-[#970747]"
             >
               {link}
             </a>
@@ -171,20 +173,19 @@ const NavBarComponent: React.FC = () => {
           <div className="border-t mt-2 pt-2 space-y-2">
             {user ? (
               <a
-                href="#"
-                className="block text-gray-800 px-4 py-2 hover:text-[#526877]"
-              >
-                Setting
-              </a>
+              
+              href='#'
+              className="block text-gray-800 px-4 py-2 hover:text-[#970747]"
+            >Setting</a>
             ) : (
               <>
                 <button
-                  className="w-full text-[#526877] border border-[#526877] px-4 py-2 rounded hover:bg-[#526877] hover:text-white"
+                  className="w-full text-[#970747] border border-[#970747] px-4 py-2 rounded hover:bg-[#970747] hover:text-white"
                   onClick={handleLogin}
                 >
                   Sign In
                 </button>
-                <button className="w-full bg-[#526877] text-white px-4 py-2 rounded hover:bg-[#526877]/90">
+                <button className="w-full bg-[#970747] text-white px-4 py-2 rounded hover:bg-[#970747]/90">
                   Sign Up
                 </button>
               </>
